+++ apps/meteor/client/views/room/MessageList/hooks/useMessageBody.tsx
@@ -1,15 +1,13 @@
 import type { IMessage } from '@rocket.chat/core-typings';
 import type { Options, Root } from '@rocket.chat/message-parser';
-import { useUserSubscription } from '@rocket.chat/ui-contexts';
 import { useMemo } from 'react';
 
 import { useAutoLinkDomains } from './useAutoLinkDomains';
-import { useAutoTranslate } from './useAutoTranslate';
+import { useMessageListAutoTranslate } from '../../../../components/message/list/MessageListContext';
 import { parseMessageTextToAstMarkdown } from '../../../../lib/parseMessageTextToAstMarkdown';
 
-export const useMessageBody = (message: IMessage | undefined, rid: string): string | Root => {
-	const subscription = useUserSubscription(rid);
-	const autoTranslateOptions = useAutoTranslate(subscription);
+export const useMessageBody = (message: IMessage | undefined): string | Root => {
+	const autoTranslateOptions = useMessageListAutoTranslate();
 	const customDomains = useAutoLinkDomains();
 
 	return useMemo(() => {