+++ apps/meteor/client/views/root/hooks/useForceLogout.ts
@@ -1,0 +1,17 @@
+import { useStream, useSessionDispatch } from '@rocket.chat/ui-contexts';
+import { useEffect } from 'react';
+
+export const useForceLogout = (userId: string) => {
+	const getNotifyUserStream = useStream('notify-user');
+	const setForceLogout = useSessionDispatch('forceLogout');
+
+	useEffect(() => {
+		setForceLogout(false);
+
+		const unsubscribe = getNotifyUserStream(`${userId}/force_logout`, () => {
+			setForceLogout(true);
+		});
+
+		return unsubscribe;
+	}, [getNotifyUserStream, setForceLogout, userId]);
+};