+++ apps/meteor/server/services/meteor/service.ts
@@ -8,7 +8,7 @@ import { Meteor } from 'meteor/meteor';
 import { MongoInternals } from 'meteor/mongo';
 
 import { triggerHandler } from '../../../app/integrations/server/lib/triggerHandler';
-import { Livechat } from '../../../app/livechat/server/lib/LivechatTyped';
+import { notifyGuestStatusChanged } from '../../../app/livechat/server/lib/guests';
 import { onlineAgents, monitorAgents } from '../../../app/livechat/server/lib/stream/agentStatus';
 import { metrics } from '../../../app/metrics/server';
 import notifications from '../../../app/notifications/server/lib/Notifications';
@@ -262,23 +262,30 @@ export class MeteorService extends ServiceClassInternal implements IMeteor {
 		return LoginServiceConfigurationModel.find({}, { projection: { secret: 0 } }).toArray();
 	}
 
-	async callMethodWithToken(userId: string, token: string, method: string, args: any[]): Promise<void | any> {
+	async callMethodWithToken(
+		userId: string,
+		token: string,
+		method: string,
+		args: any[],
+	): Promise<{
+		result: unknown;
+	}> {
 		const user = await Users.findOneByIdAndLoginHashedToken(userId, token, {
 			projection: { _id: 1 },
 		});
 		if (!user) {
 			return {
-				result: Meteor.callAsync(method, ...args),
+				result: await Meteor.callAsync(method, ...args),
 			};
 		}
 
 		return {
-			result: Meteor.runAsUser(userId, () => Meteor.callAsync(method, ...args)),
+			result: await Meteor.runAsUser(userId, () => Meteor.callAsync(method, ...args)),
 		};
 	}
 
 	async notifyGuestStatusChanged(token: string, status: UserStatus): Promise<void> {
-		return Livechat.notifyGuestStatusChanged(token, status);
+		return notifyGuestStatusChanged(token, status);
 	}
 
 	async getURL(path: string, params: Record<string, any> = {}, cloudDeepLinkUrl?: string): Promise<string> {